# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for Acomody application",
        "contact": {
            "email": "support@acomody.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/accommodation-draft": {
            "get": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Get accommodation drafts",
                "description": "Retrieves a list of accommodation drafts for the authenticated user",
                "operationId": "getAccommodationDrafts",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by draft status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "draft",
                            "enum": [
                                "draft",
                                "waiting_for_approval",
                                "published"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accommodation drafts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AccommodationDraft"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Create accommodation draft",
                "description": "Creates a new draft for an accommodation listing",
                "operationId": "createAccommodationDraft",
                "requestBody": {
                    "description": "Accommodation draft data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "description": "The draft data for the accommodation listing",
                                        "type": "object",
                                        "example": {
                                            "title": "Cozy Apartment",
                                            "description": "A nice place to stay"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Accommodation draft created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccommodationDraft"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/accommodation-draft/{accommodationDraft}": {
            "put": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Update accommodation draft",
                "description": "Updates a draft for an accommodation listing",
                "operationId": "updateAccommodationDraft",
                "parameters": [
                    {
                        "name": "accommodationDraft",
                        "in": "path",
                        "description": "Accommodation draft ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019a4b7b-3481-738a-a2ff-d93fc45bac01"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Accommodation draft data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_step",
                                    "data"
                                ],
                                "properties": {
                                    "current_step": {
                                        "description": "The current step in the accommodation listing process",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "data": {
                                        "description": "The draft data for the accommodation listing",
                                        "type": "object",
                                        "example": {
                                            "title": "Cozy Apartment",
                                            "description": "A nice place to stay"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Accommodation draft saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccommodationDraft"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/accommodation-draft/draft": {
            "get": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Get accommodation draft",
                "description": "Retrieves the saved draft for an accommodation listing",
                "operationId": "getAccommodationDraft",
                "responses": {
                    "200": {
                        "description": "Accommodation draft retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccommodationDraft"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Draft not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/accommodation-drafts/stats": {
            "get": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Get accommodation draft statistics",
                "description": "Retrieves statistics about accommodation drafts",
                "operationId": "getAccommodationDraftStats",
                "responses": {
                    "200": {
                        "description": "Accommodation draft statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Draft statistics by status",
                                            "type": "object",
                                            "example": {
                                                "draft": 3,
                                                "waiting_for_approval": 1,
                                                "published": 5
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total_drafts": {
                                                    "description": "Total number of accommodation drafts",
                                                    "type": "integer",
                                                    "example": 9
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AccommodationDraft": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "current_step": {
                        "type": "integer",
                        "example": 2
                    },
                    "status": {
                        "type": "string",
                        "example": "draft"
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "title": "Cozy Apartment",
                            "description": "A nice place to stay"
                        }
                    },
                    "last_saved_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-28T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}